Restructure project [Delivers-#163286622]
