+++ api/__init__.py
@@ -1,7 +1,8 @@
 from flask import Flask, jsonify
 import datetime
 from instance.config import app_config
-from api.views import blueprint
+from api.views.incident import blueprint
+from api.views.user import user_blueprint
 from flask_jwt_extended import JWTManager
 
 
@@ -16,6 +17,7 @@ def create_app(config_name):
 
     #  Register blueprints
     app.register_blueprint(blueprint, url_prefix='/api/v1')
+    app.register_blueprint(user_blueprint, url_prefix='/api/v1')
 
     valid_urls = [
             "GET /api/v1/",
