+++ api/models/user.py
@@ -1,0 +1,15 @@
+users = []
+
+
+class User:
+    """ A class stores user object details """
+    def __init__(self, *args):
+        self.firstname = args[0]
+        self.lastname = args[1]
+        self.othernames = args[2]
+        self.email = args[3]
+        self.phoneNumber = args[4]
+        self.username = args[5]
+        self.registered = args[6]
+        self.isAdmin = args[7]
+        self.password = args[8]
