+++ api/views/incident.py
@@ -1,17 +1,14 @@
 import datetime
 import json
 from flask import jsonify, request, Blueprint
-from api.validator import Validation, Validators
-from api.models import User, Incident
-from api.Helpers import (create_user,
-                         login_user, verify_status)
-from flask_jwt_extended import (create_access_token,
-                                jwt_required)
-from werkzeug.security import generate_password_hash, check_password_hash
+from api.validator import Validators
+from api.models.incident import Incident
+from api.Helpers import verify_status
+from flask_jwt_extended import jwt_required
 
-blueprint = Blueprint('application', __name__)
 incidents = []
-users = []
+blueprint = Blueprint('application', __name__)
+
 
 @blueprint.route('/')
 def home():
@@ -20,80 +17,6 @@ def home():
         'message': 'Welcome to bekeplar\'s iReporter app.',
         'status': '200'
     }), 200
-
-
-@blueprint.route('/signup', methods=['POST'])
-def signup():
-    """This function is used to create a new user."""
-    data = json.loads(request.data)
-    firstname = data.get('firstname')
-    lastname = data.get('lastname')
-    othernames = data.get('othernames')
-    email = data.get('email')
-    phoneNumber = data.get('phoneNumber')
-    username = data.get('username')
-    registered = datetime.datetime.utcnow()
-    isAdmin = data.get('isAdmin')
-    password = data.get('password')
-
-    user = User(firstname, lastname, othernames,
-                email, phoneNumber, username, 
-                registered, isAdmin, password
-                )
-    error = Validation.validate_input(user)
-    errors = Validation.validate_inputs(user)
-    exists = [user for user in users if user['username'] == username or user['email'] == email]
-
-    if error != None:
-        return jsonify({'Error': error}), 400
-    if errors != None:
-        return jsonify({'Error': errors}), 400
-    if exists:
-        return jsonify({
-            'message':  'user already registered.',
-            'status': 406
-            }), 406
-    password_hash = generate_password_hash(password, method='sha256')
-    create_user(username, password_hash)
-    users.append(user.__dict__)
-    return jsonify({
-        'status': 201,
-        'message': f'{username} successfully registered.',
-        'data': user.__dict__
-        }), 201
-
-
-@blueprint.route('/login', methods=['POST'])
-def login():
-    """This function allows a registered user to login"""
-    data = json.loads(request.data)
-
-    username = data.get('username')
-    password = data.get('password')
-
-    error = Validation.login_validate(username, password)
-
-    if error:
-        return jsonify({'Error': error}), 400
-    user = login_user(username, password)
-    for user in users:
-        if not user:
-            return jsonify({
-                'message': 'Wrong login credentials!',
-                'status': 401
-                }), 401
-        check_password_hash(user['password'], password) and user['username'] == username
-        access_token = create_access_token(username)
-        return jsonify({
-            'token': access_token,
-            'status': 200,
-            'message': f'{username} successfully logged in.'
-        }), 200
-    else:
-        return jsonify({
-                'message': 'Wrong login credentials!',
-                'status': 401
-                }), 401
 
 @blueprint.route('/redflags', methods=['POST'])
 @jwt_required
