+++ requirements.txt
@@ -22,6 +22,7 @@ mccabe==0.6.1
 more-itertools==4.3.0
 pep8==1.7.1
 pluggy==0.8.0
+psycopg2==2.7.6.1
 py==1.7.0
 pycodestyle==2.4.0
 pyflakes==2.0.0
